pyproject.toml
[project]
name = "streamlit-chatbot"
version = "0.1.0"
description = "Production-ready GenAI Chatbot"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "streamlit>=1.31.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0"
]

[project.optional-dependencies]
dev = [
    "ruff>=0.3.0",
    "pytest>=8.0.0",
    "black>=24.0.0"
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = []

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
